move method getContentForModel out of the Sequelize model
